refactor: #197 팀 맛집 상세 조회 시, 대기시간,준비시간,이미지 조회 되게 변경
