+++ src/main/java/com/moyorak/api/team/dto/TeamRestaurantResponse.java
@@ -1,5 +1,6 @@
 package com.moyorak.api.team.dto;
 
+import com.moyorak.api.review.domain.ReviewTimeRangeMapper;
 import com.moyorak.api.team.domain.TeamRestaurant;
 import io.swagger.v3.oas.annotations.media.Schema;
 
@@ -7,18 +8,23 @@ public record TeamRestaurantResponse(
         @Schema(description = "식당 이름", example = "김밥 천국") String name,
         @Schema(description = "한줄평", example = "여기 맛있습니다") String summary,
         @Schema(description = "장소 링크 url", example = "https://naver.me/xFLu69C9") String placeUrl,
-        @Schema(description = "음식 나오는 시간", example = "10") double servingTime,
-        @Schema(description = "대기 시간", example = "10") double waitingTime,
+        @Schema(description = "음식 나오는 시간", example = "5분이내") String servingTime,
+        @Schema(description = "대기 시간", example = "5분이내") String waitingTime,
         @Schema(description = "리뷰 갯수", example = "50") Integer reviewCount,
-        @Schema(description = "평점", example = "4.5") double score) {
-    public static TeamRestaurantResponse from(TeamRestaurant teamRestaurant) {
+        @Schema(description = "평점", example = "4.5") double score,
+        @Schema(description = "사진 경로", example = "https://..") String photoPath) {
+    public static TeamRestaurantResponse from(
+            TeamRestaurant teamRestaurant,
+            String photoPath,
+            ReviewTimeRangeMapper reviewTimeRangeMapper) {
         return new TeamRestaurantResponse(
                 teamRestaurant.getRestaurant().getName(),
                 teamRestaurant.getSummary(),
                 teamRestaurant.getRestaurant().getPlaceUrl(),
-                teamRestaurant.getAverageServingTime(),
-                teamRestaurant.getAverageWaitingTime(),
+                reviewTimeRangeMapper.mapServingTime(teamRestaurant.getAverageServingTime()),
+                reviewTimeRangeMapper.mapWaitingTime(teamRestaurant.getAverageWaitingTime()),
                 teamRestaurant.getReviewCount(),
-                teamRestaurant.getAverageReviewScore());
+                teamRestaurant.getAverageReviewScore(),
+                photoPath);
     }
 }
