+++ src/main/java/com/moyorak/api/team/service/TeamRestaurantFacade.java
@@ -1,12 +1,17 @@
 package com.moyorak.api.team.service;
 
 import com.moyorak.api.review.domain.FirstReviewPhotoPaths;
+import com.moyorak.api.review.domain.ReviewServingTime;
+import com.moyorak.api.review.domain.ReviewTimeRangeMapper;
+import com.moyorak.api.review.domain.ReviewWaitingTime;
 import com.moyorak.api.review.service.ReviewPhotoService;
+import com.moyorak.api.review.service.ReviewService;
 import com.moyorak.api.team.domain.TeamRestaurant;
 import com.moyorak.api.team.domain.TeamRestaurantSummaries;
 import com.moyorak.api.team.dto.ListResult;
 import com.moyorak.api.team.dto.TeamRestaurantListRequest;
 import com.moyorak.api.team.dto.TeamRestaurantListResponse;
+import com.moyorak.api.team.dto.TeamRestaurantResponse;
 import com.moyorak.global.domain.ListResponse;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -16,10 +21,11 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-public class TeamRestaurantListFacade {
+public class TeamRestaurantFacade {
 
     private final TeamRestaurantService teamRestaurantService;
     private final ReviewPhotoService reviewPhotoService;
+    private final ReviewService reviewService;
 
     @Transactional(readOnly = true)
     public ListResponse<TeamRestaurantListResponse> getRestaurants(
@@ -49,5 +55,25 @@ public class TeamRestaurantListFacade {
                         teamRestaurantPage.getTotalElements());
 
         return ListResponse.from(listResult.toPage(teamRestaurantListResponses));
+    }
+
+    @Transactional(readOnly = true)
+    public TeamRestaurantResponse getTeamRestaurant(
+            Long userId, Long teamId, Long teamRestaurantId) {
+        final TeamRestaurant teamRestaurant =
+                teamRestaurantService.getValidatedTeamRestaurant(teamId, teamRestaurantId);
+
+        // 조회 이벤트 발행
+        teamRestaurantService.publishTeamRestaurantEvent(userId, teamId, teamRestaurantId);
+
+        final FirstReviewPhotoPaths firstReviewPhotoPaths =
+                reviewPhotoService.findFirstReviewPhotoPaths(List.of(teamRestaurantId));
+        final String photoPath = firstReviewPhotoPaths.getPhotoPath(teamRestaurantId);
+
+        final List<ReviewServingTime> reviewServingTimes = reviewService.getAllReviewServingTimes();
+        final List<ReviewWaitingTime> reviewWaitingTimes = reviewService.getAllReviewWaitingTimes();
+        final ReviewTimeRangeMapper reviewTimeRangeMapper =
+                ReviewTimeRangeMapper.create(reviewServingTimes, reviewWaitingTimes);
+        return TeamRestaurantResponse.from(teamRestaurant, photoPath, reviewTimeRangeMapper);
     }
 }
