+++ src/test/java/com/moyorak/api/team/service/TeamRestaurantReviewFacadeTest.java
@@ -1,6 +1,7 @@
 package com.moyorak.api.team.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.SoftAssertions.assertSoftly;
 import static org.mockito.BDDMockito.given;
 
 import com.moyorak.api.restaurant.domain.Restaurant;
@@ -26,7 +27,6 @@ import com.moyorak.api.team.dto.TeamRestaurantReviewRequestFixture;
 import com.moyorak.api.team.dto.TeamRestaurantReviewResponse;
 import com.moyorak.global.domain.ListResponse;
 import java.util.List;
-import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -123,11 +123,15 @@ class TeamRestaurantReviewFacadeTest {
             assertThat(result.getData()).hasSize(1);
             final TeamRestaurantReviewResponse teamRestaurantReviewResponse =
                     result.getData().getFirst();
-            assertThat(teamRestaurantReviewResponse.id()).isEqualTo(1L);
-            assertThat(teamRestaurantReviewResponse.photoUrls())
-                    .contains("s3://review1/photo1.jpg");
-            assertThat(teamRestaurantReviewResponse.servingTime()).isEqualTo("5분이내");
-            assertThat(teamRestaurantReviewResponse.waitingTime()).isEqualTo("5분이내");
+
+            assertSoftly(
+                    it -> {
+                        it.assertThat(teamRestaurantReviewResponse.id()).isEqualTo(1L);
+                        it.assertThat(teamRestaurantReviewResponse.photoUrls())
+                                .contains("s3://review1/photo1.jpg");
+                        it.assertThat(teamRestaurantReviewResponse.servingTime()).isEqualTo("5분이내");
+                        it.assertThat(teamRestaurantReviewResponse.waitingTime()).isEqualTo("5분이내");
+                    });
         }
     }
 
@@ -164,7 +168,7 @@ class TeamRestaurantReviewFacadeTest {
                             teamId, teamRestaurantId, teamRestaurantReviewPhotoRequest);
 
             // then
-            SoftAssertions.assertSoftly(
+            assertSoftly(
                     it -> {
                         it.assertThat(result.getData()).hasSize(2);
                         it.assertThat(result.getData())
