+++ src/test/java/com/moyorak/api/team/service/TeamRestaurantServiceTest.java
@@ -18,13 +18,11 @@ import com.moyorak.api.team.domain.Team;
 import com.moyorak.api.team.domain.TeamFixture;
 import com.moyorak.api.team.domain.TeamRestaurant;
 import com.moyorak.api.team.domain.TeamRestaurantFixture;
-import com.moyorak.api.team.domain.TeamRestaurantNotFoundException;
 import com.moyorak.api.team.domain.TeamRestaurantSearch;
 import com.moyorak.api.team.domain.TeamRestaurantSearchFixture;
 import com.moyorak.api.team.domain.TeamUser;
 import com.moyorak.api.team.domain.TeamUserFixture;
 import com.moyorak.api.team.domain.TeamUserStatus;
-import com.moyorak.api.team.dto.TeamRestaurantResponse;
 import com.moyorak.api.team.dto.TeamRestaurantSaveRequest;
 import com.moyorak.api.team.dto.TeamRestaurantUpdateRequest;
 import com.moyorak.api.team.repository.TeamRestaurantRepository;
@@ -49,114 +47,6 @@ class TeamRestaurantServiceTest {
     @Mock private TeamUserRepository teamUserRepository;
     @Mock private RestaurantRepository restaurantRepository;
     @Mock private TeamRestaurantSearchRepository teamRestaurantSearchRepository;
-    @Mock private TeamRestaurantEventPublisher teamRestaurantEventPublisher;
-
-    final Long teamId = 1L;
-    final Long teamRestaurantId = 1L;
-    final Long userId = 1L;
-
-    @Nested
-    @DisplayName("팀 맛집 상세 조회 시,")
-    class GetTeamRestaurant {
-
-        @Test
-        @DisplayName("정상적으로 조회되면 응답 DTO를 반환합니다.")
-        void returnTeamRestaurantResponse() {
-            // given
-            final Restaurant restaurant =
-                    RestaurantFixture.fixture(
-                            "http://place.map.kakao.com/123456",
-                            "우가우가 차차차",
-                            "우가우가시 차차차동 24번길",
-                            "우가우가 차차로 123",
-                            RestaurantCategory.KOREAN,
-                            127.043616,
-                            37.503095);
-            final TeamRestaurant teamRestaurant =
-                    TeamRestaurantFixture.fixture(
-                            teamRestaurantId, "맛있네요", 4.5, 5, 5, 5.5, 5, true, teamId, restaurant);
-            given(teamRestaurantRepository.findByTeamIdAndIdAndUse(teamId, teamRestaurantId, true))
-                    .willReturn(Optional.of(teamRestaurant));
-
-            // when
-            final TeamRestaurantResponse response =
-                    teamRestaurantService.getTeamRestaurant(userId, teamId, teamRestaurantId);
-
-            // then
-            assertThat(response.name()).isEqualTo(teamRestaurant.getRestaurant().getName());
-        }
-
-        @Test
-        @DisplayName("팀 맛집이 존재하지 않으면 예외가 발생합니다.")
-        void throwsWhenTeamRestaurantNotFound() {
-            // given
-            given(teamRestaurantRepository.findByTeamIdAndIdAndUse(teamId, teamRestaurantId, true))
-                    .willReturn(Optional.empty());
-
-            // when & then
-            assertThatThrownBy(
-                            () ->
-                                    teamRestaurantService.getTeamRestaurant(
-                                            userId, teamId, teamRestaurantId))
-                    .isInstanceOf(TeamRestaurantNotFoundException.class);
-        }
-
-        @Test
-        @DisplayName("식당이 null이면 예외가 발생합니다.")
-        void throwsWhenRestaurantIsNull() {
-            // given
-            final TeamRestaurant teamRestaurant =
-                    TeamRestaurantFixture.fixture(
-                            teamRestaurantId, "맛있네요", 4.5, 5, 5, 5.5, 5, true, teamId, null);
-
-            given(teamRestaurantRepository.findByTeamIdAndIdAndUse(teamId, teamRestaurantId, true))
-                    .willReturn(Optional.of(teamRestaurant));
-
-            // when & then
-            assertThatThrownBy(
-                            () ->
-                                    teamRestaurantService.getTeamRestaurant(
-                                            userId, teamId, teamRestaurantId))
-                    .isInstanceOf(BusinessException.class)
-                    .hasMessage("연결된 식당 정보가 존재하지 않습니다.");
-        }
-
-        @Test
-        @DisplayName("use_yn이 'N'인 경우 조회되지 않아 예외가 발생합니다.")
-        void throwsWhenUseIsFalse() {
-            // given
-            final Restaurant restaurant =
-                    RestaurantFixture.fixture(
-                            "http://place.map.kakao.com/000000",
-                            "쓰지 않는 식당",
-                            "서울시 어디구",
-                            "서울로 456",
-                            RestaurantCategory.KOREAN,
-                            127.0,
-                            37.0);
-            final TeamRestaurant disabled =
-                    TeamRestaurantFixture.fixture(
-                            teamRestaurantId,
-                            "쓰지 않는 식당",
-                            4.0,
-                            5,
-                            5,
-                            5.0,
-                            5,
-                            false,
-                            teamId,
-                            restaurant);
-            given(teamRestaurantRepository.findByTeamIdAndIdAndUse(teamId, teamRestaurantId, true))
-                    .willReturn(Optional.empty());
-
-            // when & then
-            assertThatThrownBy(
-                            () ->
-                                    teamRestaurantService.getTeamRestaurant(
-                                            userId, teamId, teamRestaurantId))
-                    .isInstanceOf(TeamRestaurantNotFoundException.class);
-        }
-    }
 
     @Nested
     @DisplayName("팀 맛집 저장 시")
